Rename: BridgeMaker, BridgeRandomNumberGenerator 파일 위치 변경
