+++ docs/README.md
@@ -2,39 +2,39 @@
 
 ## 기능 요구사항
 
-1.  [ ] 다리 길이를 입력받는다.
-2.  [ ] 다리 길이만큼 0과 1이 랜덤하게 섞인 다리를 생성한다.
+1.  [x] 다리 길이를 입력받는다.
+2.  [x] 다리 길이만큼 0과 1이 랜덤하게 섞인 다리를 생성한다.
 
     - 무작위 값이 0인경우 아래로 이동 가능 1인 경우 위로 이동 가능하다.
     -     다리길이가 4, 생성된 랜덤 값이 [1,1,1,1] 인 경우
           [ O | O | O | O ]
           [   |   |   |   ] 가 정답이 된다.
 
-3.  [ ] 이동할 칸을 입력 받는다.
+3.  [x] 이동할 칸을 입력 받는다.
 
     - 이동할 칸이 윗다리면 U 아랫다리면 D를 입력받는다.
     - U가 1 D가 0이라 생각하고 각각의 현재의 idx값과 비교해 볼것.
 
-4.  [ ] 입력 값이 이동 가능한 경우 3번으로 돌아간다
-5.  [ ] 입력 값이 이동 불가능한 경우 재시작/게임 종료 여부를 물어본다.
-    1.  [ ] 재시작(R)을 입력했을 경우 view초기화 후 3번으로 돌아간다
-    2.  [ ] 게임 종료(Q)를 입력했을 경우 지금까지 시도 횟수와 성공 여부를 나타낸다.
-6.  [ ] 다리 끝에 도달했을 경우 총 시도 횟수와 성공 여부를 나타낸다.
+4.  [x] 입력 값이 이동 가능한 경우 3번으로 돌아간다
+5.  [x] 입력 값이 이동 불가능한 경우 재시작/게임 종료 여부를 물어본다.
+    1.  [x] 재시작(R)을 입력했을 경우 view초기화 후 3번으로 돌아간다
+    2.  [x] 게임 종료(Q)를 입력했을 경우 지금까지 시도 횟수와 성공 여부를 나타낸다.
+6.  [x] 다리 끝에 도달했을 경우 총 시도 횟수와 성공 여부를 나타낸다.
 
 ---
 
 ### 예외 처리 사항
 
-1. [ ] 다리 길이 입력
+1. [x] 다리 길이 입력
 
    - 숫자만 입력 가능합니다.(타입)
    - 3이상 20이하 입력 가능합니다.(범위)
 
-2. [ ] 방향 입력 (U,D)
+2. [x] 방향 입력 (U,D)
 
    - U,D만 입력 가능합니다.
 
-3. [ ] 게임 재시작 입력 (R,Q)
+3. [x] 게임 재시작 입력 (R,Q)
    - R,Q만 입력 가능합니다.
 
 ## 프로그래밍 요구 사항
@@ -42,7 +42,7 @@
 - [ ] indent(인덴트, 들여쓰기) depth를 3이 넘지 않도록 구현한다. 2까지만 허용한다.
 - [ ] 함수(또는 메서드)가 한 가지 일만 하도록 최대한 작게 만들어라.
 - [ ] Jest를 이용하여 본인이 정리한 기능 목록이 정상 동작함을 테스트 코드로 확인한다.
-- [ ] else를 지양한다.
+- [x] else를 지양한다.
   - 힌트: if 조건절에서 값을 return하는 방식으로 구현하면 else를 사용하지 않아도 된다.
   - 때로는 if/else, switch문을 사용하는 것이 더 깔끔해 보일 수 있다. 어느 경우에 쓰는 것이 적절할지 스스로 고민해 본다.
 - [ ] 도메인 로직에 단위 테스트를 구현해야 한다. 단, UI(Console.readLine, Console.print) 로직에 대한 단위 테스트는 제외한다.
@@ -51,8 +51,8 @@
 ---
 
 - [ ] 함수(또는 메서드)의 길이가 10라인을 넘어가지 않도록 구현한다.
-- [ ] 메서드의 파라미터 개수는 최대 3개까지만 허용한다.
-- [ ] 아래 있는 InputView, OutputView, BridgeGame, BridgeMaker 클래스(또는 객체)의 요구사항을 참고하여 구현한다.
+- [x] 메서드의 파라미터 개수는 최대 3개까지만 허용한다.
+- [x] 아래 있는 InputView, OutputView, BridgeGame, BridgeMaker 클래스(또는 객체)의 요구사항을 참고하여 구현한다.
   - InputView 에서만 MissionUtils의 Console.readLine() 을 이용해 사용자의 입력을 받을 수 있다.
   - BridgeGame 클래스에서 InputView, OutputView 를 사용하지 않는다.
 
@@ -82,8 +82,8 @@
 
 ## 과제 진행 요구사항
 
-- [ ] 미션은 javascript-bridge 저장소를 Fork & Clone해 시작한다.
-- [ ] 기능을 구현하기 전 docs/README.md에 구현할 기능 목록을 정리해 추가한다.
-- [ ] Git의 커밋 단위는 앞 단계에서 docs/README.md에 정리한 기능 목록 단위로 추가한다.
-  - [ ]커밋 메시지 컨벤션 가이드를 참고해 커밋 메시지를 작성한다.
-- [ ] 과제 진행 및 제출 방법은 프리코스 과제 제출 문서를 참고한다.
+- [x] 미션은 javascript-bridge 저장소를 Fork & Clone해 시작한다.
+- [x] 기능을 구현하기 전 docs/README.md에 구현할 기능 목록을 정리해 추가한다.
+- [x] Git의 커밋 단위는 앞 단계에서 docs/README.md에 정리한 기능 목록 단위로 추가한다.
+  - [x]커밋 메시지 컨벤션 가이드를 참고해 커밋 메시지를 작성한다.
+- [x] 과제 진행 및 제출 방법은 프리코스 과제 제출 문서를 참고한다.
