+++ src/utils/BridgeMaker.js
@@ -1,7 +1,7 @@
 /**
  * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
  */
-const { DEFAULT } = require("./utils/constant.js");
+const { DEFAULT } = require("./constant.js");
 const BridgeMaker = {
   /**
    * @param {number} size 다리의 길이
